_contrib: add bash one-liner to look up C-Lightning PRs
C-Lightning rebases commits onto its master branch rather than merging
and so doesn't add any information about which PR originated a commit in
its commit log.  However, GitHub provides that information in its web
interface, so this command makes it possible to look up a C-Lightning
commit from the command line.

It's pretty specific to Debian/Ubuntu-based Linux distros as-is, but
since it's really just one line of code, anyone else who needs to use it
should be able to port it as needed.