[build] Don't jekyll clean for previews
The Makefile for this project was copied from bitcoin.org, where
incremental builds didn't work because of their custom plugins.
Therefore, the build was cleaned for every preview.

We don't have complicated plugins so we can just do an incremental build
for local previews.

Also add a `make clean` target.