Tests: allow only retesting changed topic and compat YAML
Previously, we ran the JSON schema validator on all YAML files for the
Compatibility Matrix and Topic Pages every time we ran the pre-build
tests.  Unfortunately, the schema validator is pretty slow and the
increasing number of files checked by it is creating a delay.

With this update, we only check all the files on a fresh checkout, a
build after a clean, or on a `make production` build.  Otherwise, we
only check changed files.

This edit also fixes concurrency, so you can safely run with `make -j4`
or whatever.  This doesn't give you much speedup, but maybe that can be
improved in the future.