Podcast: simplify newsletter references
Removes the need for boilerplate logic when adding newsletter references
The new logic requires just the podcast timestamps to be added in the
related newsletter sections.
For each podcast, the `recap_references_generator` will parse the
`reference` newsletter (declared in frontmatter) looking for podcast
reference marks (timestamps) in order to create the newsletter
references for the podcast page.