Topics: anchor-outputs fix wrong statement
An unconfirmed commitment transaction does not allow funds to be stolen.
The timelock in the `to_local` output (see BOLT03) is constructed with an OP_CSV
which makes it relative to the height of the tx confirmation.
The situation actually leads to funds stuck from the point of view of
the party that started the unilateral close. He cannot CPFP this
transaction due to the relative timelock.