Use CommonMark block markdown style for Kramdown markdown=1
Kramdown allows an HTML block or span element to contain markdown=1,
markdown=span, or markdown=block to trigger markdown parsing within that
element.  Hugo's Goldmark Commonmark parser requires an empty space
after the HTML element open and before the element close to trigger the
same behavior.

This commit adds the Commonmark behavior in addition to the Kramdown
behavior.  We'll have to maintain both until we're fully transitioned.

No test for this is provided.